Change name of files and extract custom settings to other file.
